tools/bili-analyzer/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "bili-analyzer"
version = "0.1.0"
description = "Stable Bilibili video prepare pipeline (yt-dlp + ffmpeg + dedup)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "Pillow>=10.0.0",
  "mcp>=1.0.0"
]

[project.optional-dependencies]
asr = [
  "faster-whisper>=1.0.0"
]

[project.scripts]
bili-analyzer = "bili_analyzer.cli:main"
bili-analyzer-mcp = "bili_analyzer.mcp_server:run"

[tool.setuptools]
packages = ["bili_analyzer"]
